modify and delete some components
