+++ src/utils/index.tsx
@@ -1,7 +1,10 @@
 import React from 'react'
-import { IconContext } from 'react-icons'
-import { FaTwitter } from 'react-icons/fa'
-import { MdCreate, MdSearch, MdHome, MdLocalOffer, MdMoreHoriz, MdClose } from 'react-icons/md'
+import { parseISO, format } from 'date-fns'
+
+export function Date({ dateString }: { dateString: string }) {
+  const date = parseISO(dateString)
+  return <time dateTime={dateString}>{format(date, 'LLLL d, yyyy')}</time>
+}
 
 export function NoImage (){
   return (
@@ -58,65 +61,4 @@ export const WantedlySvg = (props) => {
 
 export const HatenaSvg = (props) => {
   return <svg className={props.class} xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24'><path d='M20.47 0C22.42 0 24 1.58 24 3.53v16.94c0 1.95-1.58 3.53-3.53 3.53H3.53C1.58 24 0 22.42 0 20.47V3.53C0 1.58 1.58 0 3.53 0h16.94zm-3.705 14.47c-.78 0-1.41.63-1.41 1.41s.63 1.414 1.41 1.414 1.41-.645 1.41-1.425-.63-1.41-1.41-1.41zM8.61 17.247c1.2 0 2.056-.042 2.58-.12.526-.084.976-.222 1.32-.412.45-.232.78-.564 1.02-.99s.36-.915.36-1.48c0-.78-.21-1.403-.63-1.87-.42-.48-.99-.734-1.74-.794.66-.18 1.156-.45 1.456-.81.315-.344.465-.824.465-1.424 0-.48-.103-.885-.3-1.26-.21-.36-.493-.645-.883-.87-.345-.195-.735-.315-1.215-.405-.464-.074-1.29-.12-2.474-.12H5.654v10.486H8.61zm.736-4.185c.705 0 1.185.088 1.44.262.27.18.39.495.39.93 0 .405-.135.69-.42.855-.27.18-.765.254-1.44.254H8.31v-2.297h1.05zm8.656.706v-7.06h-2.46v7.06H18zM8.925 9.08c.71 0 1.185.08 1.432.24.245.16.367.435.367.83 0 .38-.13.646-.39.804-.265.154-.747.232-1.452.232h-.57V9.08h.615z' /></svg>
-}
-
-// icons
-export const MoreIcon = () => {
-  return (
-    <React.Fragment>
-      <IconContext.Provider value={{ className: 'react-icons' }}><MdMoreHoriz /></IconContext.Provider>
-      <style jsx global>{`
-        .react-icons, .homeIconSvg{
-          position: absolute;
-          top: 50%;
-          left: 50%;
-          transform: translate(-50%, -50%);
-          fill: #424242;
-          width: 1rem;
-          height: 1rem;
-        }
-
-        @media ( min-width: 960px ){
-          .react-icons, .homeIconSvg{
-            width: 1.75rem;
-            height: 1.75rem;
-          }
-        }
-      `}</style>
-    </React.Fragment>
-  )
-}
-
-export const CloseIcon = () => {
-  return (
-    <IconContext.Provider value={{ className: 'react-icons' }}><MdClose /></IconContext.Provider>
-  )
-}
-
-export const SearchIcon = () => {
-  return (
-    <IconContext.Provider value={{ className: 'react-icons' }}><MdSearch /></IconContext.Provider>
-  )
-}
-
-export const HomeIcon = () => {
-  return (
-    <IconContext.Provider value={{ className: 'react-icons' }}><MdHome /></IconContext.Provider>
-  )
-}
-
-export const PostsIcon = () => {
-  return (
-    <IconContext.Provider value={{ className: 'react-icons' }}><MdCreate /></IconContext.Provider>
-  )
-}
-export const TagsIcon = () => {
-  return (
-    <IconContext.Provider value={{ className: 'react-icons' }}><MdLocalOffer /></IconContext.Provider>
-  )
-}
-export const TwitterIcon = () => {
-  return (
-    <IconContext.Provider value={{ className: 'react-icons' }}><FaTwitter /></IconContext.Provider>
-  )
 }
